Implement player damage and enemy interaction mechanics
- Add player damage handling in backend game state
- Introduce DamagedIntent client message for enemy interactions
- Update server message types to support damage confirmation
- Modify enemy and player tag types to use smaller integer ranges
- Refactor collision and game flow components to support new damage system

+++ game/src/flow/online_game/trigger/spawn_enemy.rs
@@ -5,7 +5,7 @@ use crate::components::{Velocity, UFO};
 
 #[derive(Event)]
 pub struct SpawnEnemyEvent {
-    pub tag: u128,
+    pub tag: u16,
     pub position: (f32, f32),
     pub velocity: (f32, f32),
 }
